CodeAI/requirements.txt
opencv-python
paho-mqtt
streamlit
pandas
watchdog